print first 50 fibonacci's

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-	int x = 0;
-	int y = 1;
+	long int x = 0;
+	long int y = 1;
 	int i;
-	int s = x + y;
+	long int s = x + y;
 
 	for (i = 3; i < 53; ++i)
 	{
